Reset settings if source settings not there
In the copysettingsgroup function:
If a specific setting bundle is null or empty, this certainly mean that the standard settings have been set for the corresponding group. The settings of the target should be reseted rather than remained without changes.

I guess that the proofing of "targetlanguage" is necessary but I haven't tested it.