- Created a new service "AccessTokenService" to automatically recover the authorization "Access token" that is now added to the header for call (search) to the IATE terminology service.
The service will also identify when the token as expired and attempt to extend (by way of refreshing) the token for an additional 9hrs.
Note: If the total time allowed for the access token as elapsed (i.e. > 12 hours) and the project has been open for that same period of time, then an exception is thrown indicating an error in refreshing the access token.