Added null validation when the provider is added from Translate single document. (#835)
Save Requires secure connection option and populate the ui when the plugin is added to another project.
Removed the validation when not all language pairs receives dictionary from Edge (in case we add the provider form Translate single document and we receive more language pairs). Put back tne Throw exception in Factory if the credentials were not found in Credential Store.