Updates:     Enable user to override the default system paths for MyDocument & ApplicationData from a configuration file.     Removed 'Sdl.Community.' from the assembly output names.
Details:
Starting with this latest release, a new configuration file 'Qualitivity.dll.config' is unpacked in plugins folder (...\Plugins\Unpacked\Qualitivity\Qualitivity.dll.config) when you load the plugin in Trados Studio. The configuration file enables the user to override the default system paths for MyDocumentsPath & ApplicationDataPath
Note: It is possible to specify the same folder for both.  If the folder/s specified are empty or not valid, then the default system path/s are used.

Example: (Qualitivity.dll.config)

<?xml version="1.0" encoding="utf-8" ?>
<configuration>
 <appSettings>
  <add key="MyDocumentsPath" value="C:\Temp\" />
  <add key="ApplicationDataPath" value="C:\Temp\" />
 </appSettings>
</configuration>